# backend/docs/swagger/swagger.yaml
basePath: /
definitions:
  blotting-consultancy_internal_model.JSONMap:
    additionalProperties: true
    type: object
...
  internal_handler_content.GetDraftResponse:
    properties:
      config:
        $ref: '#/definitions/blotting-consultancy_internal_model.JSONMap'
      pageKey:
        type: string
      publishedVersion:
        type: integer
      updatedAt:
        type: string
      version:
        type: integer
    type: object
  internal_handler_content.GetVersionDetailResponse:
    properties:
      config:
        $ref: '#/definitions/blotting-consultancy_internal_model.JSONMap'
      createdBy:
        type: integer
      id:
        type: integer
      pageKey:
        type: string
      publishedAt:
        type: string
      version:
        type: integer
    type: object
...
  internal_handler_content.UpdateDraftRequest:
    properties:
      changeNote:
        type: string
      config:
        $ref: '#/definitions/blotting-consultancy_internal_model.JSONMap'
    required:
    - config
    type: object
  internal_handler_content.UpdateDraftResponse:
    properties:
      pageKey:
        type: string
      updatedAt:
        type: string
      version:
        type: integer
    type: object
...
host: localhost:8088
info:
  contact: {}
  description: Bilingual CMS backend API for Impress (印迹). Supports content management,
    articles, pages, themes, media, and more.
  title: Impress CMS API
  version: "1.0"
paths:
...
  /admin/categories:
    get:
      description: Returns all categories for admin management
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              items:
                items:
                  type: object
                type: array
            type: object
      security:
      - BearerAuth: []
      summary: List all categories
      tags:
      - Categories (Admin)
    post:
      consumes:
      - application/json
      description: Create a new category
      parameters:
      - description: Category data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - Categories
  /admin/categories/{id}:
    delete:
      description: Delete a category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete category
      tags:
      - Categories
    get:
      description: Returns a single category by its database ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get category by ID
      tags:
      - Categories (Admin)
    put:
      consumes:
      - application/json
      description: Update an existing category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated category data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - Categories
  /admin/categories/tree:
    get:
      description: Returns categories organized in a parent-child tree structure
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              items:
                items:
                  type: object
                type: array
            type: object
      security:
      - BearerAuth: []
      summary: List categories as tree
      tags:
      - Categories (Admin)
...
  /admin/content/{pageKey}/draft:
    get:
      description: Returns the draft config for a given page key
      parameters:
      - description: Page key (e.g. home, about)
        in: path
        name: pageKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_content.GetDraftResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get draft content
      tags:
      - Content (Admin)
    put:
      consumes:
      - application/json
      description: Updates the draft config for a page key with optimistic locking
        via If-Match header
      parameters:
      - description: Page key
        in: path
        name: pageKey
        required: true
        type: string
      - description: Expected draft version for optimistic locking
        in: header
        name: If-Match
        required: true
        type: integer
      - description: Draft content data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler_content.UpdateDraftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_content.UpdateDraftResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update draft content
      tags:
      - Content (Admin)
...
  /admin/content/{pageKey}/versions/{version}:
    get:
      description: Returns the config snapshot for a specific published version
      parameters:
      - description: Page key
        in: path
        name: pageKey
        required: true
        type: string
      - description: Version number
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_content.GetVersionDetailResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get version detail
      tags:
      - Content (Admin)
...
  /public/categories/{slug}:
    get:
      description: Returns a category by slug with paginated published articles
      parameters:
      - description: Category slug
        in: path
        name: slug
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              articles:
                type: object
              category:
                type: object
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get category by slug
      tags:
      - Categories
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" for JWT authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"